refactor: create separate folder for controllers for a neat filestructure
